Initial refactor of ResourceForm
